Build fix for Ubuntu.
QtMultimedia has been moved to QtMobility in Ubuntu and as a result CMake cannot find the library anymore. Hopefully this works better in Qt5, but in the meanwhile compile the Qt based audio output only if the multimedia library has been found. If not found, use what FreeRDP provides for audio output.

Signed-off-by: Janne Hakonen <janne.hakonen@oss.tieto.com>